This file includes the code for the spam prediction of the total dataset using DST
This code works with the WEBSPAM-UK2007. It's main purpose is to find relative likelihood of a host of being spam for finally  classifying it as spam or non-spam. However, application of Dempster-Shafer Theory allows non-precise probability of a host being spam or benign. The presented model is not sensitive to small changes to the belief assignment and changes do not alter the operating characteristics of the proposed mathematical model.